separate preprocessing
